PF-41 implement mobile nav flow
