+++ src/common/links/side-nav-link/index.jsx
@@ -3,15 +3,23 @@ import cx from "classnames";
 import { NavLink as RouterNavLink } from "react-router-dom";
 
 import { Icon } from "common/icon";
+import { Avatar } from "common/avatar";
 
 import "./styles.scss";
+
+const Variant = {
+  default: "default",
+  profile: "profile",
+};
 
 export const SideNavLink = ({
   to,
   icon,
   title,
   onClick,
+  height = 56,
   isCompact = false,
+  variant = Variant.default,
 }) => {
   const linkClassName = useCallback(
     (isActive) =>
@@ -26,10 +34,12 @@ export const SideNavLink = ({
     <RouterNavLink
       to={to}
       onClick={onClick}
+      style={{ height }}
       className={(isActive) => linkClassName(isActive)}
     >
       <div className="icon-container">
-        <Icon icon={icon} />
+        {variant === Variant.default && <Icon icon={icon} />}
+        {variant === Variant.profile && <Avatar placeholderIcon="person" />}
       </div>
       {title}
     </RouterNavLink>
