+++ src/common/page-header/change-password-modal/config.js
