+++ src/common/page-header/change-password-modal/styles.scss
