+++ src/common/page-header/settings-modal/styles.scss
