+++ src/common/page-header/styles.scss
@@ -48,8 +48,6 @@
   @include breakpoint("tablet", "max") {
     height: 72px;
     padding: 0px 16px;
-
-    z-index: 10;
     position: relative;
 
     .right-nav-container {
