+++ src/common/side-bar-navigation/calendar/date-header/styles.scss
