+++ src/common/side-bar-navigation/calendar/helpers.js
