+++ src/common/side-bar-navigation/calendar/styles.scss
