+++ src/common/side-bar-navigation/expand-button/index.jsx
