+++ src/common/side-bar-navigation/expand-button/styles.scss
