+++ src/common/side-bar-navigation/styles.scss
