+++ src/screens/calendarScreen/CalendarScreen.jsx
@@ -13,7 +13,7 @@ import {
   currYearName,
   currMonthName,
   buildCalendar,
-} from "common/navigation/side-navigation/calendar/helpers";
+} from "common/side-bar-navigation/calendar/helpers";
 
 import useSizeFinder from "../../utils/sizeFinder";
 import Button from "../../components/button/Button";
